requirements.txt
discord.py==2.7.1
ollama==0.6.1
python-dotenv==1.2.2